refactor: 💡 move tooltippopup to common folder
Authors: Marcus Hogler, Gustav Pettersson
