+++ src/common/TooltipPopup/TooltipContent/TooltipContent.types.ts
