+++ src/common/TooltipPopup/TooltipPopup.types.ts
@@ -1,17 +1,19 @@
 import { ColorFn } from 'common/Types';
 import React from 'react';
-import { Props as TooltipProps } from '../Tooltip.types';
+
+export type InModal = boolean;
+export type Position = 'top' | 'right' | 'bottom' | 'left';
 
 export type Props = {
   ref?: React.RefObject<any>;
   label: React.ReactNode;
-  position: TooltipProps['position'];
-  positionCallback?: (calulatedPostion: NonNullable<TooltipProps['position']>) => void;
+  position?: Position;
+  positionCallback?: (calulatedPostion: Position) => void;
   triggerElement?: Element;
   ariaLabel?: string;
   backgroundColor?: ColorFn;
   borderColor?: ColorFn;
-  inModal?: TooltipProps['inModal'];
+  inModal?: InModal;
   maxWidth?: number;
   pointerEvents?: boolean;
 };
