+++ src/common/TooltipPopup/consts.ts
@@ -1,0 +1,3 @@
+// Values below is in px
+export const BORDER_SIZE = 1;
+export const TRIANGLE_SIZE = 10;
